requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.2
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.3.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.5.4
dataclasses-json==0.6.7
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
email_validator==2.2.0
exceptiongroup==1.2.1
Faker==25.9.2
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.12
langchain-text-splitters==0.2.2
langsmith==0.1.82
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.35.7
orjson==3.10.5
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pylint==3.2.3
pytest==8.2.2
pytest-alembic==0.11.0
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tenacity==8.4.2
tiktoken==0.7.0
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
